+++ lib/src/commands/has_git_references.dart
@@ -31,6 +31,7 @@ class HasGitReferences extends DirCommand<bool> {
   Future<bool> exec({
     required Directory directory,
     required GgLog ggLog,
+    Map<String, dynamic> options = const {},
   }) async {
     return await GgStatusPrinter<bool>(
       message: 'Checking for git references.',