+++ lib/src/commands/is_ahead_main.dart
@@ -29,6 +29,7 @@ class IsAheadMain extends DirCommand<bool> {
   Future<bool> exec({
     required Directory directory,
     required GgLog ggLog,
+    Map<String, dynamic> options = const {},
   }) async {
     return await GgStatusPrinter<bool>(
       message: 'Checking if branch is ahead main.',