+++ lib/src/commands/update_project_git.dart
@@ -27,6 +27,7 @@ class UpdateProjectGit extends DirCommand<bool> {
   Future<bool> exec({
     required Directory directory,
     required GgLog ggLog,
+    Map<String, dynamic> options = const {},
   }) async {
     return await GgStatusPrinter<bool>(
       message: 'Updating Git branches.',